Refactor multidevice allocation utilities
