+++ csrc/multidevice/utils.h
@@ -7,8 +7,6 @@
 // clang-format on
 #pragma once
 
-#include <c10/util/ArrayRef.h>
-
 #include <compute_at_map.h>
 #include <fusion.h>
 #include <ir/interface_nodes.h>
@@ -16,11 +14,6 @@
 #include <visibility.h>
 
 namespace nvfuser {
-
-// Return true if the TensorView is contiguous. This function is more
-// permissive than torch.Tensor.is_contiguous because it allows expanded
-// broadcasts.
-bool isTvContiguous(const TensorView* tv);
 
 // Returns whether a TensorView has a non-reduction axis parallelized Didx
 // Checks that the other non-reduction axis are not parallelized on Didx
@@ -115,20 +108,5 @@ std::unordered_map<int64_t, int64_t> reorderParallelizedToFront(TensorView*);
 // Validate the expression is a valid DID split: expr is an outer split with
 // device dim as the outer dimension.
 bool isValidDeviceSplit(Expr* expr);
-
-// Find the producing logical id of the given allocation id traversing
-// through device splits. For unsharded allocation_id, logical_id is the same as
-// allocation_id.
-IterDomain* projectShardedAllocationToLogical(
-    TensorView* tv,
-    IterDomain* allocation_id);
-
-// Finds the allocated id corresponding to the given logical id
-// traversing through device splits. For e.g.: `i0` -> `DIDx(d), i0/d` will
-// return `i0/d`. For unsharded logical_id, allocation_id is the same as
-// logical_id.
-IterDomain* projectLogicalToShardedAllocation(
-    TensorView* tv,
-    IterDomain* logical_id);
 
 } // namespace nvfuser
