+++ csrc/preseg_passes/finalize_multidevice_domains.cpp
@@ -13,6 +13,7 @@
 #include <ir/iostream.h>
 #include <ir/utils.h>
 #include <linked_hash_map.h>
+#include <multidevice/allocation_utils.h>
 #include <multidevice/utils.h>
 #include <scheduler/utils.h>
