+++ csrc/scheduler/utils.cpp
@@ -25,6 +25,7 @@
 #include <ir/interface_nodes.h>
 #include <ir/utils.h>
 #include <logical_domain_map.h>
+#include <multidevice/allocation_utils.h>
 #include <multidevice/utils.h>
 #include <ops/all_ops.h>
 #include <scheduler/mma_utils.h>
